Reorganize serialization code in preparation for larger changes. (#1107)
