+++ src/sst/core/serialization/CMakeLists.txt
@@ -14,30 +14,16 @@
 # ~~~
 #
 
+add_subdirectory(impl)
+
 set(SSTSerializationHeaders
-    serializable_fwd.h
-    serializable_base.h
     serializable.h
-    serialize_array.h
-    serialize_atomic.h
-    serialize_buffer_accessor.h
-    serialize_deque.h
+    serializable_base.h
     serialize.h
     serialize_impl_fwd.h
-    serialize_list.h
-    serialize_map.h
-    serialize_multiset.h
-    serialize_output.h
-    serialize_packer.h
-    serialize_priority_queue.h
-    serializer_fwd.h
-    serializer.h
     serialize_serializable.h
-    serialize_set.h
-    serialize_sizer.h
-    serialize_string.h
-    serialize_unpacker.h
-    serialize_vector.h
+    serializer.h
+    serializer_fwd.h
     statics.h)
 
 install(FILES ${SSTSerializationHeaders}
