+++ src/sst/core/serialization/impl/serialize_deque.h
@@ -9,43 +9,36 @@
 // information, see the LICENSE file in the top level directory of the
 // distribution.
 
-#ifndef SST_CORE_SERIALIZATION_SERIALIZE_PRIORITY_QUEUE_H
-#define SST_CORE_SERIALIZATION_SERIALIZE_PRIORITY_QUEUE_H
+#ifndef SST_CORE_SERIALIZATION_IMPL_SERIALIZE_DEQUE_H
+#define SST_CORE_SERIALIZATION_IMPL_SERIALIZE_DEQUE_H
+
+#ifndef SST_INCLUDING_SERIALIZE_H
+#warning \
+    "The header file sst/core/serialization/impl/serialize_deque.h should not be directly included as it is not part of the stable public API.  The file is included in sst/core/serialization/serialize.h"
+#endif
 
 #include "sst/core/serialization/serializer.h"
 
-#include <queue>
+#include <deque>
 
 namespace SST {
 namespace Core {
 namespace Serialization {
 
-
-template <class T, class S, class C>
-class serialize<std::priority_queue<T, S, C>>
+template <class T>
+class serialize<std::deque<T>>
 {
-    typedef std::priority_queue<T, S, C> Pqueue;
+    typedef std::deque<T> Deque;
 
 public:
-    S& getContainer(std::priority_queue<T, S, C>& q)
-    {
-        struct UnderlyingContainer : std::priority_queue<T, S, C>
-        {
-            static S& getUnderlyingContainer(std::priority_queue<T, S, C>& q) { return q.*&UnderlyingContainer::c; }
-        };
-        return UnderlyingContainer::getUnderlyingContainer(q);
-    }
-
-    void operator()(Pqueue& v, serializer& ser)
+    void operator()(Deque& v, serializer& ser)
     {
         switch ( ser.mode() ) {
         case serializer::SIZER:
         {
             size_t size = v.size();
             ser.size(size);
-
-            auto container = getContainer(v);
-            for ( auto it = container.begin(); it != container.end(); ++it ) {
+            for ( auto it = v.begin(); it != v.end(); ++it ) {
                 T&   t = const_cast<T&>(*it);
                 ser& t;
             }
@@ -55,9 +48,7 @@ public:
         {
             size_t size = v.size();
             ser.pack(size);
-
-            auto container = getContainer(v);
-            for ( auto it = container.begin(); it != container.end(); ++it ) {
+            for ( auto it = v.begin(); it != v.end(); ++it ) {
                 T&   t = const_cast<T&>(*it);
                 ser& t;
             }
@@ -70,7 +61,7 @@ public:
             for ( size_t i = 0; i < size; ++i ) {
                 T    t = {};
                 ser& t;
-                v.push(t);
+                v.push_back(t);
             }
             break;
         }
@@ -82,4 +73,4 @@ public:
 } // namespace Core
 } // namespace SST
 
-#endif // SST_CORE_SERIALIZATION_SERIALIZE_PRIORITY_QUEUE_H
+#endif // SST_CORE_SERIALIZATION_IMPL_SERIALIZE_DEQUE_H
