+++ src/sst/core/serialization/impl/unpacker.h
@@ -9,10 +9,15 @@
 // information, see the LICENSE file in the top level directory of the
 // distribution.
 
-#ifndef SST_CORE_SERIALIZATION_SERIALIZE_UNPACKER_H
-#define SST_CORE_SERIALIZATION_SERIALIZE_UNPACKER_H
+#ifndef SST_CORE_SERIALIZATION_IMPL_UNPACKER_H
+#define SST_CORE_SERIALIZATION_IMPL_UNPACKER_H
 
-#include "sst/core/serialization/serialize_buffer_accessor.h"
+#ifndef SST_INCLUDING_SERIALIZER_H
+#warning \
+    "The header file sst/core/serialization/impl/unpacker.h should not be directly included as it is not part of the stable public API.  The file is included in sst/core/serialization/serializer.h"
+#endif
+
+#include "sst/core/serialization/impl/ser_buffer_accessor.h"
 
 namespace SST {
 namespace Core {
@@ -44,4 +49,4 @@ public:
 } // namespace Core
 } // namespace SST
 
-#endif // SST_CORE_SERIALIZATION_SERIALIZE_UNPACKER_H
+#endif // SST_CORE_SERIALIZATION_IMPL_UNPACKER_H
