--- src/sst/core/serialization/serializable_fwd.h
@@ -1,25 +1,0 @@
-// Copyright 2009-2024 NTESS. Under the terms
-// of Contract DE-NA0003525 with NTESS, the U.S.
-// Government retains certain rights in this software.
-//
-// Copyright (c) 2009-2024, NTESS
-// All rights reserved.
-//
-// This file is part of the SST software package. For license
-// information, see the LICENSE file in the top level directory of the
-// distribution.
-
-#ifndef SST_CORE_SERIALIZATION_SERIALIZABLE_FWD_H
-#define SST_CORE_SERIALIZATION_SERIALIZABLE_FWD_H
-
-namespace SST {
-namespace Core {
-namespace Serialization {
-
-class serializable;
-
-} // namespace Serialization
-} // namespace Core
-} // namespace SST
-
-#endif // SST_CORE_SERIALIZATION_SERIALIZABLE_FWD_H
